Moved dir to config pkg
